requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
pymysql>=1.1.0
python-dotenv>=1.0.0
pandas>=2.2.3,<3.0.0
numpy>=1.26.0,<2.0.0
requests>=2.31.0
torch>=2.1.0,<2.3.0
einops>=0.6.0
china_stock_data==0.2.1
safetensors>=0.3.0

# Optional (uncomment only if needed):
# plotly==5.17.0
# huggingface_hub>=0.34.0,<1.0
# transformers>=4.35.0,<5.0.0
# datasets>=2.14.0,<3.0.0
